Feature/ti 86 save portfolio to follow in a table in bdd

+++ smartimmo/src/main/java/com/gangdestrois/smartimmo/domain/buyer/BuyerManager.java
@@ -19,7 +19,7 @@ public class BuyerManager implements BuyerApi {
     }
 
     @Override
-    public Buyer findBuyerById(int id) {
+    public Buyer findBuyerById(Long id) {
         return buyerSpi.findBuyerById(id);
     }
 }
